% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_nas.R
\name{analyze_nas}
\alias{analyze_nas}
\title{Analyze and produce information on NAs in dataset.}
\usage{
analyze_nas(
  df,
  output_filename = "NA_Proportions.csv",
  na_file = TRUE,
  full_table_print = FALSE
)
}
\arguments{
\item{df}{The data frame.}

\item{output_filename}{The filename for the written csv file, containing the
table of NA information. Default is "NA_Proportions.csv". Note: leaving
this default will cause the file to overwrite when run more than once in
the same directory.}

\item{na_file}{When set to TRUE, this will write a csv file with
the table of NA counts and proportions. If set to FALSE, it will print
the table but not generate a csv file.}

\item{full_table_print}{TRUE or FALSE, defaults to FALSE. If FALSE, the NAs
table will only print the head of the resulting table. If TRUE, the NAs
table will print the entire table of NAs.}
}
\value{
A data frame with columns Column_Index, Column_Name, NA_Count,
NA_Proportion. The rows consist of the columns in the inputted data frame
which contain NAs. If there are no NAs in the inputted data frame, an
empty data frame is returned. Additionally, a csv is written to the
current working directory with the same information.
}
\description{
This function examines the inputted dataset for NA values and outputs a table
of counts and proportions of NAs for each appropriate variable in a data
frame within the R console and a generated csv file, written to the
current working directory.
}
