% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detox.R
\name{detox}
\alias{detox}
\title{Detox the data! Clean and reorganize an RSA-911 dataset.}
\usage{
detox(
  data,
  NA_check = TRUE,
  unidentified_to_0 = FALSE,
  remove_strictly_NA = FALSE
)
}
\arguments{
\item{data}{An RSA-911 dataset.}

\item{NA_check}{TRUE or FALSE, defaults to TRUE. If TRUE, Runs analyze_nas
function on dataset after cleaning to output information on remaining NA
values.}

\item{remove_strictly_NA}{TRUE or FALSE, defaults to FALSE. If TRUE,
.   identifies and removes variables with exclusively NA values after the data
.   have been cleaned.}

\item{unidentified_to_O}{TRUE or FALSE, defaults to FALSE. If TRUE, converts
.   values of 9 to 0 within variables with possible values of 0, 1, 9 (where 9
.   represents "did not identify").}
}
\value{
The cleaned dataset (additionally, a table of NA information if
NA_check param is set to TRUE. See analyze_nas documentation for info.)
}
\description{
This function is a one-stop for cleaning and restructuring an RSA-911 dataset
.   in order to prepare it for analysis and visualization. Processed include
.   tailored NA imputation, variable type conversion, and reorganization of
.   variables.
}
