% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{py_20_sample}
\alias{py_20_sample}
\title{A Random Subset of RSA-911 PY2020exit}
\format{
\subsection{\code{py_20_sample}}{

A data frame with 300 rows and 307 columns:
\describe{
\item{variables}{For a list of all 307 variables, check the website.}
...
}
}
}
\source{
\url{https://rsa.ed.gov/performance/rsa-911-policy-directive}
}
\usage{
py_20_sample
}
\description{
A subset of data from the RSA-911 PY2020exit dataset
}
\keyword{datasets}
